E2Utility 1.2
1. Removed all Essential Orbwalker combo checkers to deteremine the user is combo'ing or not. Now the assigned key under Activator configs is the only one to check the user is combo'ing.

Thanks to @Silktouch

2. Added minimum CC duration checkers to Cleansers.

[SPOILER]

795

[/SPOILER]

3. Added Flee option on Cleanses

I have no idea why I didn't add it on the initial release, but while I'm implementing minimum CC duration for Cleanses, I found out I didn't add flee.

4. Now consumables will not be used in the ally fountain.

Thanks to @ketoth

5. Shared XP Detector (Beta) is added.

[SPOILER]

 796

[/SPOILER]

A reason why I call it as "Beta" is still it is kind of buggy.

Not like it's causing crashes but "rarely" it calculates wrongly the number of enemies around. (I would say like 1 out of 30-40)

Most of the time, it works fine for now. I have some ideas to improve it, but I need some time to work on.

And I need some feedback, so I'm releasing it as beta.

6. Custom Ranges is Added

[SPOILER]

799

[/SPOILER]

It is a pretty simple feature.

I think no one is going to use it.

But I added it because I am going to use it.

It saves different configs on every champion.

The good thing of it is you can check ranges settings such as Gank Alerter range, Shared xp range to know how far actually they are.